Removed Dependency on resource_dasm
Implemented the first method to extract the image from the resource fork then rebuild it using xxd. had to reconvert it using `sips` otherwise it seems to not work with app bundles.
Remving the bytes at the end of the file didnt help as well.
endig number should be `454E 44AE 4260 82`

Signed-off-by: asamahy <amrashraf@protonmail.com>